+++ src/lib.rs
@@ -5,7 +5,7 @@
 pub mod error;
 pub mod file_ops;
 pub mod passphrase;
-pub mod secretcrypt;
+pub mod secretcrypt_v1;
 pub mod varmor;
 
 pub use error::{ErrorCategory, ErrorKind, Result, SaltyboxError};
